Fixed minor inconsistencies and added a debug feature to app.py. Just type "debug" in the first input, and it should give a quick way of testing if the notes are being played appropriately. The app.spec file is useful for making executables with pyinstaller.

+++ source/notes.py
@@ -17,7 +17,7 @@ AUDIO_FOLDER = 'aiff'
 import platform
 os_name = platform.system()
 
-def bpmToSeconds(bpm: int) -> float:
+def bpmToSeconds(bpm: float) -> float:
     return 60 / bpm
 
 def check_ffmpeg():
